# server/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Self-hosted note-taking application API. Health probes are available at root paths `/livez` and `/readyz` (outside `/api/v1`).",
        "title": "Jot API",
        "contact": {
            "name": "Jot Project",
            "url": "https://github.com/hanzei/jot"
        },
        "license": {
            "name": "MIT"
        },
        "version": "1.0"
    },
    "basePath": "/api/v1",
    "paths": {
...
        "/pats": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "pats"
                ],
                "summary": "List personal access tokens for the current user",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handlers.PATListResponse"
                        }
                    },
                    "401": {
                        "description": "unauthorized",
                        "schema": {
                            "$ref": "#/definitions/apierr.ErrorResponse"
                        }
                    }
                },
                "security": [
                    {
                        "CookieAuth": []
                    }
                ]
            },
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "pats"
                ],
                "summary": "Create a new personal access token",
                "parameters": [
                    {
                        "description": "Token name",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handlers.createPATRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/handlers.patResponse"
                        }
                    },
                    "400": {
                        "description": "bad request",
                        "schema": {
                            "$ref": "#/definitions/apierr.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "unauthorized",
                        "schema": {
                            "$ref": "#/definitions/apierr.ErrorResponse"
                        }
                    },
                    "422": {
                        "description": "personal access token cap exceeded (pat_limit_reached)",
                        "schema": {
                            "$ref": "#/definitions/apierr.ErrorResponse"
                        }
                    }
                },
                "security": [
                    {
                        "CookieAuth": []
                    }
                ]
            }
        },
        "/pats/{id}": {
            "delete": {
                "tags": [
                    "pats"
                ],
                "summary": "Revoke a personal access token",
                "parameters": [
                    {
                        "type": "string",
                        "description": "PAT ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "no content"
                    },
                    "401": {
                        "description": "unauthorized",
                        "schema": {
                            "$ref": "#/definitions/apierr.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "not found",
                        "schema": {
                            "$ref": "#/definitions/apierr.ErrorResponse"
                        }
                    }
                },
                "security": [
                    {
                        "CookieAuth": []
                    }
                ]
            }
        },
...
    },
    "definitions": {
        "apierr.Code": {
            "type": "string",
            "enum": [
                "validation_failed",
                "unauthorized",
                "forbidden",
                "not_found",
                "method_not_allowed",
                "conflict",
                "request_too_large",
                "limit_exceeded",
                "rate_limited",
                "bad_request",
                "internal",
                "invalid_credentials",
                "session_required",
                "registration_disabled",
                "local_login_disabled",
                "incorrect_password",
                "username_taken",
                "cannot_share_with_self",
                "already_shared",
                "label_name_taken",
                "last_admin",
                "cannot_delete_self",
                "sso_identity_linked",
                "would_strand_account",
                "label_name_too_long",
                "item_limit_reached",
                "image_limit_reached",
                "pat_limit_reached",
                "unsupported_image_type",
                "invalid_image",
                "invalid_import_file",
                "sso_link_unavailable",
                "sso_unlink_unavailable"
            ],
            "x-enum-comments": {
                "CodeAlreadyShared": "409 — the note is already shared with that user.",
                "CodeBadRequest": "any other 4xx.",
                "CodeCannotDeleteSelf": "403 — admins cannot delete their own account.",
                "CodeCannotShareWithSelf": "400 — a note cannot be shared with its owner.",
                "CodeConflict": "409 — the request conflicts with current state (duplicate ID, concurrent modification, …).",
                "CodeForbidden": "403 — authenticated, but not allowed to do this.",
                "CodeImageLimitReached": "422 — the note would exceed the per-note image cap.",
                "CodeIncorrectPassword": "403 — the current password given to change a password is wrong.",
                "CodeInternal": "any 5xx. The message is always masked.",
                "CodeInvalidCredentials": "401 from login — wrong username or password.",
                "CodeInvalidImage": "400 — the uploaded file claims an accepted type but does not decode as an image.",
                "CodeInvalidImportFile": "400 — the import file is not a readable export of the chosen import type.",
                "CodeItemLimitReached": "422 — the note would exceed the per-note item cap.",
                "CodeLabelNameTaken": "400 from rename, 409 from a create with a client-supplied ID — the user already has a label with this name.",
                "CodeLabelNameTooLong": "422 — the label name exceeds the label name length limit.",
                "CodeLastAdmin": "409 — the change would leave the server without an admin.",
                "CodeLimitExceeded": "422 — a resource cap or length limit was exceeded.",
                "CodeLocalLoginDisabled": "403 from register, login, and password change — password login is turned off; use SSO.",
                "CodeMethodNotAllowed": "405 — the route exists but not for this method.",
                "CodeNotFound": "404 — the resource or route does not exist (or is not visible to the caller).",
                "CodePATLimitReached": "422 — the user already holds the maximum number of personal access tokens.",
                "CodeRateLimited": "429 — too many requests; honor Retry-After.",
                "CodeRegistrationDisabled": "403 — self-registration is turned off.",
                "CodeRequestTooLarge": "413 — the request body exceeds the endpoint's limit.",
                "CodeSSOIdentityLinked": "409 — the SSO identity belongs to another account.",
                "CodeSSOLinkUnavailable": "403 — password login is turned off, so there is no local account to link SSO to.",
                "CodeSSOUnlinkUnavailable": "403 — password login is turned off, so SSO is the only way to sign in and cannot be disconnected.",
                "CodeSessionRequired": "403 — the endpoint needs a browser session, and the request authenticated with a personal access token.",
                "CodeUnauthorized": "401 — missing, invalid, or expired credentials.",
                "CodeUnsupportedImageType": "400 — the uploaded file is not one of the endpoint's accepted image types.",
                "CodeUsernameTaken": "409 — another account already has this username.",
                "CodeValidationFailed": "400 — malformed or invalid request input.",
                "CodeWouldStrandAccount": "422 — unlinking SSO would leave the account with no way to sign in."
            },
            "x-enum-descriptions": [
                "400 — malformed or invalid request input.",
                "401 — missing, invalid, or expired credentials.",
                "403 — authenticated, but not allowed to do this.",
                "404 — the resource or route does not exist (or is not visible to the caller).",
                "405 — the route exists but not for this method.",
                "409 — the request conflicts with current state (duplicate ID, concurrent modification, …).",
                "413 — the request body exceeds the endpoint's limit.",
                "422 — a resource cap or length limit was exceeded.",
                "429 — too many requests; honor Retry-After.",
                "any other 4xx.",
                "any 5xx. The message is always masked.",
                "401 from login — wrong username or password.",
                "403 — the endpoint needs a browser session, and the request authenticated with a personal access token.",
                "403 — self-registration is turned off.",
                "403 from register, login, and password change — password login is turned off; use SSO.",
                "403 — the current password given to change a password is wrong.",
                "409 — another account already has this username.",
                "400 — a note cannot be shared with its owner.",
                "409 — the note is already shared with that user.",
                "400 from rename, 409 from a create with a client-supplied ID — the user already has a label with this name.",
                "409 — the change would leave the server without an admin.",
                "403 — admins cannot delete their own account.",
                "409 — the SSO identity belongs to another account.",
                "422 — unlinking SSO would leave the account with no way to sign in.",
                "422 — the label name exceeds the label name length limit.",
                "422 — the note would exceed the per-note item cap.",
                "422 — the note would exceed the per-note image cap.",
                "422 — the user already holds the maximum number of personal access tokens.",
                "400 — the uploaded file is not one of the endpoint's accepted image types.",
                "400 — the uploaded file claims an accepted type but does not decode as an image.",
                "400 — the import file is not a readable export of the chosen import type.",
                "403 — password login is turned off, so there is no local account to link SSO to.",
                "403 — password login is turned off, so SSO is the only way to sign in and cannot be disconnected."
            ],
            "x-enum-varnames": [
                "CodeValidationFailed",
                "CodeUnauthorized",
                "CodeForbidden",
                "CodeNotFound",
                "CodeMethodNotAllowed",
                "CodeConflict",
                "CodeRequestTooLarge",
                "CodeLimitExceeded",
                "CodeRateLimited",
                "CodeBadRequest",
                "CodeInternal",
                "CodeInvalidCredentials",
                "CodeSessionRequired",
                "CodeRegistrationDisabled",
                "CodeLocalLoginDisabled",
                "CodeIncorrectPassword",
                "CodeUsernameTaken",
                "CodeCannotShareWithSelf",
                "CodeAlreadyShared",
                "CodeLabelNameTaken",
                "CodeLastAdmin",
                "CodeCannotDeleteSelf",
                "CodeSSOIdentityLinked",
                "CodeWouldStrandAccount",
                "CodeLabelNameTooLong",
                "CodeItemLimitReached",
                "CodeImageLimitReached",
                "CodePATLimitReached",
                "CodeUnsupportedImageType",
                "CodeInvalidImage",
                "CodeInvalidImportFile",
                "CodeSSOLinkUnavailable",
                "CodeSSOUnlinkUnavailable"
            ]
        },
        "apierr.ErrorDetail": {
            "type": "object",
            "properties": {
                "code": {
                    "description": "Code is stable and machine-readable; clients switch on it.",
                    "allOf": [
                        {
                            "$ref": "#/definitions/apierr.Code"
                        }
                    ],
                    "example": "not_found"
                },
                "message": {
                    "description": "Message is human-readable English, for logs and as a fallback. Masked\nto \"internal server error\" for 5xx responses.",
                    "type": "string",
                    "example": "note not found"
                }
            }
        },
        "apierr.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "$ref": "#/definitions/apierr.ErrorDetail"
                }
            }
        },
...
        "handlers.PATListResponse": {
            "type": "object",
            "properties": {
                "pats": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/handlers.patResponse"
                    }
                }
            }
        },
...
        "handlers.createPATRequest": {
            "type": "object",
            "properties": {
                "name": {
                    "type": "string"
                }
            }
        },
        "handlers.patResponse": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "token": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "CookieAuth": {
            "type": "apiKey",
            "name": "jot_session",
            "in": "header"
        }
    }
}